requirements.txt
mss
Pillow
openai
python-dotenv
PySide6
